pyproject.toml
[project]
name = "apibooks"
version = "0.1.0"
description = ""
authors = []
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "pydantic[email] (>=2.11.7,<3.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "asyncpg (>=0.30.0,<0.31.0)"
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
alembic = "^1.16.5"
pytest-asyncio = "^1.1.0"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
